% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L0hima.R
\name{L0hima}
\alias{L0hima}
\title{L0-regularized high-dimensional mediation analysis}
\usage{
L0hima(
  X,
  M,
  Y,
  COV.XM = NULL,
  COV.MY = NULL,
  family = c("gaussian", "binomial", "poisson", "cox", "AFT"),
  test.type = c("JS-uniform", "JS-mixture"),
  test.control = c("FWER", "FDR"),
  sig.level = 0.05,
  screening = FALSE,
  topN = NULL
)
}
\arguments{
\item{X}{A vector of exposure.}

\item{M}{a \code{matrix} of mediators. Rows represent samples, columns represent variables.}

\item{Y}{The outcome variable, of \code{n} observations.
For \code{family = "gaussian"} \code{Y} should be a vector of continious outcome.
For \code{family = "binomial"} \code{Y} should be a vector of binomial outcome.
For \code{family = "poisson"} \code{Y} should be a vector of count outcome.
For \code{family = "cox"} or \code{"AFT"}, \code{Y} should be a two-column matrix with columns
named \code{"time"} and \code{"status"}.}

\item{COV.XM}{a \code{matrix} of covariates dataset for testing the association \code{X ~ M}.
Covariates specified here will not participate penalization.}

\item{COV.MY}{a \code{matrix} of covariates dataset for testing the association \code{M ~ Y}.
Covariates specified here will not participate penalization.}

\item{family}{One of the following models:
\code{"gaussian"} (continuous outcome),
\code{"binomial"} (binomial outcome),
\code{"poisson"} (count outcome),
\code{"cox"} (right-censored outcome),
\code{"AFT"} (right-censored outcome).}

\item{test.type}{The type of significant test. Available options are \code{"JS-uniform"}, and \code{"JS-mixture"}.}

\item{test.control}{The objective of significant test. Available options are \code{"FWER"}, and \code{"FDR"}.}

\item{sig.level}{The significance level of test. Default is 0.05.}

\item{screening}{Whether to conduct the sure independent screening. Default = \code{TRUE}.}

\item{topN}{An integer specifying the number of top markers from sure independent screening. Default = \code{NULL}.}
}
\value{
A data.frame containing results of L0-regularized high-dimensional mediation analysis.
\itemize{
    \item{SM: }{The index set of selected mediators.}
    \item{A_est: }{estimates of coefficient A (predictor (X) --> mediators (M)).}
    \item{B_est: }{estimates of coefficient B (mediators (M) --> outcome (Y)).}
}
}
\description{
\code{L0hima} is used to estimate and test high-dimensional mediation effects.
}
\examples{
n <- 400  # sample size
p <- 1000 # the dimension of mediators
q <- 2 # the number of covariates
family <- "gaussian"  # the type of model

# the regression coefficients A (X --> M)
A<-rep(0,p)

# the regression coefficients B (M --> Y)
B<-rep(0,p)

# the first four markers are true mediators.
A[1:6] <- c(0.45, 0.4, -0.35, -0.3, 0.0, 0.5)
B[1:6] <- c(0.5, -0.45, 0.4, -0.35, 0.5,0.0)
rhoM<-0.0 #the correlation coefficient of random errors for M
# Generate simulation data
dat = simL0hima(n,p,q,family,A,B,rhoM,seed=1234)
L0hima.fit<-L0hima(dat$X, dat$M, dat$Y, dat$Z, dat$Z, family,
test.type="JS-uniform",test.control="FDR", sig.level=0.05 )
L0hima.fit
}
