% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simL0hima.R
\name{simL0hima}
\alias{simL0hima}
\title{Simulation data generator for L0-regularized high-dimensional mediation analysis}
\usage{
simL0hima(
  n,
  p,
  q,
  family = c("gaussian", "binomial", "poisson", "cox", "AFT"),
  A,
  B,
  rhoM = 0,
  seed = 1000
)
}
\arguments{
\item{n}{An integer specifying sample size.}

\item{p}{An integer specifying the dimension of mediators.}

\item{q}{An integer specifying the dimension of covariates.}

\item{family}{One of the following models:
\code{"gaussian"} (continuous outcome),
\code{"binomial"} (binomial outcome),
\code{"poisson"} (count outcome),
\code{"cox"} (right-censored outcome),
\code{"AFT"} (right-censored outcome).}

\item{A}{An vector of the regression coefficients A \code{X ~ M}.}

\item{B}{An vector of the regression coefficients B \code{M ~ Y}.}

\item{rhoM}{An coefficient between 0 and 1, which determines the correlation matrix of random errors for M.}

\item{seed}{An integer specifying a seed for random number generation.}
}
\value{
A list containing the information about the generated data
\itemize{
    \item{X: }{A vector of predictor.}
    \item{M: }{A matrix of mediators. Rows represent samples, columns represent variables.}
    \item{Y: }{The outcome variable.}
    \item{Z: }{A matrix of covariates.}
    \item{A: }{An vector of the regression coefficients A \code{X ~ M}.}
    \item{B: }{An vector of the regression coefficients B \code{M ~ Y}.}
}
}
\description{
\code{simL0hima} is used to generate simulation data for L0-regularized high-dimensional mediation analysis.
}
\examples{
n <- 400  # sample size
p <- 1000 # the dimension of mediators
q <- 2 # the number of covariates
family <- "gaussian"  # the type of model

# the regression coefficients A (X --> M)
A<-rep(0,p)

# the regression coefficients B (M --> Y)
B<-rep(0,p)

# the first four markers are true mediators.
A[1:6] <- c(0.45, 0.4, -0.35, -0.3, 0.0, 0.5)
B[1:6] <- c(0.5, -0.45, 0.4, -0.35, 0.5,0.0)
rhoM<-0.0 #the correlation coefficient of random errors for M
# Generate simulation data
dat = simL0hima(n,p,q,family,A,B,rhoM,seed=1234)
}
